Transferred from the Adobe Open Source Subversion Repository
Revision: 16727
Author:   jszeto@adobe.com
Date:     2010-06-29T20:42:03.768739Z

Log Message: 
-----------
Form and ComboBox Bug Fixes

SDK-25765 - ComboBox does not commit item when closed if Enter key is pressed

I changed the fix for this bug.

Instead of only committing the selection in ComboBox.applySelection if the dropDown is closed, we keep track of the user defined selection in DropDownList.keyDownHandler. 

SDK-26268 - ComboBox does not commit value with TAB when openOnInput = false

Removed the restriction in ComboBox.applySelection that the dropDown must be open in order to commit

SDK-26442 -  When arrowing between the first two items in a flex combo box the selection is always on the secodn item

Fixed logic bug in ComboBox.changeHighlightedSelection. 

SDK-26752 - error string no longer appears for Spark or MX components

The errorString was getting set in MXML. At that time, the component is not parented and not hooked up to its styles. So the call to get the showErrorTip style was returning false. 

I moved this logic to commitProperties which solves the timing issue. 

SDK-26497 - Spark Form displays errors in a seemingly random order

Changed the dictionary of invalid elements into a sorted array. 
An element's sort order is based on its position in the display list tree using a depth first search. The position is represented by a Vector.<int>, with each item being the child index of the element's ancestor at that depth. For example, the second child of the first child of the root would have a Vector like this [0,0,1]. 

Whenever an element is invalid, we add it to the sorted array and perform a re-sort. We can live with using a heavy data structure like a Vector of ints because there should only be a few errors at a time in a Form. The alternative is to traverse the entire tree and store a pointer and the sort order for each node. 

We make an assumption that the order in the display list tree is equivalent to the visual order of the Form's input controls. Since the Form's layout is based on VerticalLayout, this is a valid assumption. This same algorithm would work if the Form was switched to a horizontal layout. 

SDK-26697 - Cannot reset Spark Form to original state if form had validation errors

Added logic to Form to listen for the errorStringChanged event on an invalid element. We don't use the same technique as in FormItem, which listens for errorStringChanged in the capture phase on the FormItem's contentGroup. This is because FormItem's errorString is set to the combined errorStrings of its children. If Form used the same technique, the error strings would appear twice. 

QE notes: retest SDK-25765
Doc notes: none
Bugs: SDK-25765, SDK-26268, SDK-26442, SDK-26752, SDK-26497, SDK-26697
Reviewer: Kevin
Tests run: full run
Is noteworthy for integration: no

Modified Paths:
--------------
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/framework/src/mx/core/UIComponent.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/ComboBox.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/supportClasses/DropDownListBase.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/Form.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1259979 13f79535-47bb-0310-9956-ffa450edef68