Transferred from the Adobe Open Source Subversion Repository
Revision: 6368
Author:   gosmith@adobe.com
Date:     2009-04-29T02:20:06.064038Z

Log Message: 
-----------
Undid typo fix in SWFBridgeEvent ("brdigeWindowDeactivate" -> "bridgeWindowDeactivate") that I made in revision 6343.

This class is involved in the "Marshall Plan" feature of making SWFs compiled against different versions of the framework work together. Therefore any change has to be made across Flex 3 and Flex 4. And since versions of Flex 3 with the misspelling have already shipped, it needs to be made in a way such that apps with the misspelled event can communicate with apps with the correctly spelled event. So I'll file a bug and leave it to Darrell or Alex if we decide to fix it.

QE Notes: None
Doc Notes: None
Bugs: None
Reviewer: Darrell and Alex

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/framework/src/mx/events/SWFBridgeEvent.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1260761 13f79535-47bb-0310-9956-ffa450edef68