Transferred from the Adobe Open Source Subversion Repository
Revision: 11423
Author:   dloverin@adobe.com
Date:     2009-11-04T13:37:42.534490Z

Log Message: 
-----------
Style manager fixes

StyleManagerImpl:
The code to remove style declarations from Subject was calling delete to remove the declaration. It removed the declaration but left an undefined entry in the Array because the array length was not changed. Later when the parent and child arrays are concatenated an undefined entry exists which causes the exception. Using splice() keeps the array length correct. Additionally, when there is only one entry in the array the entire Object from _subject is removed.

Alert.show() is called from a module and crashes because it is trying to use styles in the top-level app but the top-level app does not use Alert so the styles are not loaded. A work around would be to declare Alert in the top level application to draw in the needed styles. I made a fix in Alert so that work around is not needed.

Panel:
Null pointer check for titleBarBackground. This fix avoids the RTE but the Alert has no background.

Alert:
Add a module factory argument to Alert.show(). This is an alternative workaround to declaring Alert in the top-level application.


QE notes: None.
Doc notes: None.
Bugs: SDK-23977, SDK-23974, SDK-23995
Reviewer: Alex
Tests run: checkintests, components/Alert, Managers/StyleManager, Managers/PopUpManager
Is noteworthy for integration: No.

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/styles/StyleManagerImpl.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/containers/Panel.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/controls/Alert.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1267698 13f79535-47bb-0310-9956-ffa450edef68