Transferred from the Adobe Open Source Subversion Repository
Revision: 9328
Author:   rfrishbe@adobe.com
Date:     2009-08-15T00:02:35.436187Z

Log Message: 
-----------
Fixing SDK-22462: Event dispatched is not as expected in Tree

For this, the issue is really that TreeItemRenderers assume that no click event can occur on the disclosureIcon.  This is because the discolsureIcon is added and removed each time it goes through invalidation in commitProperties(), so a new discolsureIcon gets created and replaces the old one after they mouseDown (meaning no click event actually occurs).  Automation replays some mouse events on one disclosureIcon only (which is an automation bug because that’s not how it actually works in the real world) and when a “click” occurs, that causes an itemClick on the Tree, when it shouldn’t.  For Flex 3, this wasn’t a problem because the IR was getting invalidated through a different codepath because openDuration did not default to 0.  If you set it to 0, you’d see the same problem in Flex 3.  Anyways, the fix is just to block all click events that occur on the disclosureIcon.

QE notes: -
Doc notes: -
Bugs: SDK-22462
Reviewer: Alex
Tests run: checkintests, mustella Tree
Is noteworthy for integration: No

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/framework/src/mx/controls/treeClasses/TreeItemRenderer.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1269125 13f79535-47bb-0310-9956-ffa450edef68