Transferred from the Adobe Open Source Subversion Repository
Revision: 3048
Author:   jvroom@adobe.com
Date:     2008-08-29T22:40:17.623182Z

Log Message: 
-----------
Always map urls through the rootURL property, even for the non-proxy case.  This might
change the behavior slightly if you set rootURL on HTTPService for the non-proxy case
as that would have been ignored in the current code path but it seem less consistent.
If rootURL is not set, it defaults to the swf's URL which should give the same relative
URL semantics we get today by default when rootURL is not set.

blazeDS checkintests pass

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/rpc/src/mx/rpc/http/AbstractOperation.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1274159 13f79535-47bb-0310-9956-ffa450edef68