Transferred from the Adobe Open Source Subversion Repository
Revision: 3732
Author:   jvroom@adobe.com
Date:     2008-10-17T23:23:59.708467Z

Log Message: 
-----------
Fix for HTTPMultiService.showBusyCursor property.   The AbstractOperation class was
setting the property in its constructor.   The Operation class though is keying off of
the property being set at the operation level and ignores the service level setting
in that case.  So by always setting the property at the operation level, the service
level setting never took effect.

Thanks to Sameer for finding the bug, Seth for the review.

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/rpc/src/mx/rpc/http/AbstractOperation.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1274163 13f79535-47bb-0310-9956-ffa450edef68