Transferred from the Adobe Open Source Subversion Repository
Revision: 10848
Author:   gosmith@adobe.com
Date:     2009-10-02T22:59:52.046845Z

Log Message: 
-----------
Removed the [ExcludeClass] metadata on the out-of-package class FTETextFieldStyleResolver in TLFTextField.as. It was actually causing TLFTextField to be excluded, and is not necessary on an out-of-package class, which is invisible except to code in the same file.

QE notes: None
Doc notes: None
Bugs: None
Reviewer: Carol
Tests run: ant checkintests
Is noteworthy for integration: No

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/mx/core/TLFTextField.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1277883 13f79535-47bb-0310-9956-ffa450edef68