Transferred from the Adobe Open Source Subversion Repository
Revision: 12027
Author:   gosmith@adobe.com
Date:     2009-11-20T05:43:31.265390Z

Log Message: 
-----------
Added selectionAnchorIndex and selectionActiveIndex getters to RichEditableTextAccImpl and SkinnableTextBaseAccImpl. The presence of these getters is supposed to make Player 10.1 and AIR 2.0 able to support text-selection-accessibility for TLF-based text; they are the ActionScript side of the GetSelection() method in Adobe's ISimpleTextSelection COM interface for screen readers like JAWS.

Although Player 10.1 and AIR 2.0 are adding a flash.accessibility.ISimpleTextSelection interface to formalize these getters, we want to be able to compile against the playerglobal.swc for Player 10.0 (or airglobal.swc for AIR 1.5). We could define our own copy of this interface in Flex, but it isn't necessary because the Player and AIR do not actually look at whether an AccessibilityImplementation class declares that it implements ISimpleTextSelection; they only look at whether an AccessibilityImplementation class has these getters.

This way of doing text-selection-accessibility replaces the ITextSupport interface that was in earlier versions of Player 10.1 and AIR 2.0, including the betas. So you try it you must have a very recent build of 10.1 and 2.0.

QE notes: None
Doc notes: None
Bugs: None
Reviewer: None
Tests run: None
Is noteworthy for integration: No

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/spark/src/spark/accessibility/RichEditableTextAccImpl.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/spark/src/spark/accessibility/SkinnableTextBaseAccImpl.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1278068 13f79535-47bb-0310-9956-ffa450edef68