Transferred from the Adobe Open Source Subversion Repository
Revision: 5004
Author:   rfrishbe@adobe.com
Date:     2009-02-19T01:17:22.291072Z

Log Message: 
-----------
Working on cleaning up the renderers again when the dataProvider is set.  At the dataProvider setter time, we clean up the old ones.  At commitProperties() time we create the new ones (in the virtual case, this is done at updateDisplayList() time).

I had to add some cleanup around a dataProvider reset or refresh because we need to do this cleanup there as well.  However, we can't rely on the dataProvider to loop through the item renderers because it may have changed, so we rely on our own data structure, indexToRenderers.  This makes finding the actual item associated with the renderer a little hard, but for most use-cases it should be fine.

QE Notes: -
Doc Notes: -
Bugs: SDK-19364
Reviewer: Glenn
tests: checkintests, DataGroup, FxList, FxButtonBar, FxDataContainer

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/DataGroup.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1278677 13f79535-47bb-0310-9956-ffa450edef68