Transferred from the Adobe Open Source Subversion Repository
Revision: 5917
Author:   jszeto@adobe.com
Date:     2009-04-03T21:28:26.428083Z

Log Message: 
-----------
Re-implementation of DropDownList

DropDownList now subclasses List instead of DropDownBase (thanks to Iwo Banas!). 
Dropdown base functionality has been moved up to DropDownList. The skin now contains the the same elements as the ListSkin. 
The component is in a usable and stable state. The one major issue is that the selectedIndex changes immediately in response to user interaction. The correct behavior is for selectedIndex to change when the dropdown closes. 

QE Notes: DropDownList has a new base class
Doc Notes: DropDownList has a new base class. DropDownBase is removed.
Bugs: n/a
Reviewer: Glenn

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/spark/components/DropDownList.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/spark/components/List.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/spark/skins/default/DropDownListSkin.mxml</path>
<path kind="file" action="D">/flex/sdk/trunk/frameworks/projects/flex4/src/spark/components/supportClasses/DropDownBase.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1278796 13f79535-47bb-0310-9956-ffa450edef68